Added support for metadata for job runs
With the `laminarc tag` command the build scripts can augment runs with
metadata. This can be SCM-oriented information like the used branch or
source-version.
The metadata is a list of key/value pairs. The metadata is automatically
shown in the run-view of a job in the web UI.

Each call of `laminarc tag` can assign one key/value pair.

Example:
```
laminarc tag $JOBNAME $NUMBER` \
		Branch "$(git -C $WORKSPACE branch --show-current)"
```